requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
black==23.10.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docopt==0.6.2
drf-nested-routers==0.93.4
executing==2.0.0
idna==3.4
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
model-bakery==1.17.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==1.7.1
PyMySQL==1.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-watch==4.2.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.7
watchdog==3.0.0
wcwidth==0.2.8
